backend/requirements.txt
flask
Flask-PyMongo
python-dotenv
passlib
Flask-CORS
bcrypt
PyJWT
Flask-reuploaded
gunicorn
boto3